+++ app/src/main/java/com/zgld/mall/beans/OrderItems.java
@@ -12,7 +12,7 @@ public class OrderItems extends AbstractOrderItems implements java.io.Serializab
 	}
 
 	/** full constructor */
-	public OrderItems(Integer orderId, Integer productId, Double sku, Integer quantity, Double listPrice, Double cellPrice, String remark) {
+	public OrderItems(Integer orderId, Integer productId, Integer sku, Integer quantity, Double listPrice, Double cellPrice, String remark) {
 		super(orderId, productId, sku, quantity, listPrice, cellPrice, remark);
 	}
